+++ spec/controllers/teams_controller_spec.rb
@@ -13,6 +13,19 @@ RSpec.describe TeamsController, type: :controller do
   let(:team) { create(:team, owners: [owner]) }
 
   describe 'GET #show' do
+
+    it 'paginates team users' do
+      sign_in owner
+      get :show, id: team.id
+      expect(assigns(:team_users)).to respond_to(:total_pages)
+    end
+
+    it 'paginates namespaces' do
+      sign_in owner
+      get :show, id: team.id
+      expect(assigns(:team_namespaces)).to respond_to(:total_pages)
+    end
+
     it 'allows team members to view the page' do
       sign_in owner
       get :show, id: team.id