+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Upcoming Version
+
+- Teams page is now paginated
+- Teams namespaces and users are now paginated in team detail page
+
 ## 1.0.1
 
 - Fixed regression where namespaces could not be created from team page